[#342] refactor: ClaimReport 형태로 리팩토링
