+++ src/main/java/com/jeju/nanaland/domain/report/entity/claim/ReportType.java
@@ -1,0 +1,5 @@
+package com.jeju.nanaland.domain.report.entity.claim;
+
+public enum ReportType {
+  MEMBER, REVIEW
+}
