+++ src/main/java/com/jeju/nanaland/domain/report/repository/ClaimReportRepository.java
@@ -1,0 +1,13 @@
+package com.jeju.nanaland.domain.report.repository;
+
+import com.jeju.nanaland.domain.member.entity.Member;
+import com.jeju.nanaland.domain.report.entity.claim.ClaimReport;
+import com.jeju.nanaland.domain.report.entity.claim.ReportType;
+import java.util.Optional;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface ClaimReportRepository extends JpaRepository<ClaimReport, Long> {
+
+  Optional<ClaimReport> findByMemberAndIdAndReportType(Member member, Long id,
+      ReportType reportType);
+}
