+++ src/main/java/com/jeju/nanaland/domain/report/repository/ClaimReportVideoFileRepository.java
@@ -1,0 +1,9 @@
+package com.jeju.nanaland.domain.report.repository;
+
+import com.jeju.nanaland.domain.report.entity.claim.ClaimReportVideoFile;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface ClaimReportVideoFileRepository extends
+    JpaRepository<ClaimReportVideoFile, Long> {
+
+}
